Add YouTube Equi-Angular Cubemap (EAC) projection
YouTube has been using this projection since 2018 (see
ytdl-org/youtube-dl#15267), and most YouTube downloaders will get
EAC format when attempting to retrieve YouTube 360° videos. This
patch adds ability to render videos in this projection. Formulas
for the shader are derived from Google's blog [1], and pixel
truncation done by trial and error with a bit of pixel inspection
on a single frame of a downloaded video. Without turncating the
pixels, seams appear on cube borders where the textures are
discontinuous.

Two videos downloaded from YouTube, in both EAC projection and
Equirectangular projection, has been added to the repo, along with
relevant HTMLs to demonstrate the rendering. During the testing I
found that split Equirectangular projection to not render at all,
and that has been fixed along with this patch.

[1] https://blog.google/products/google-ar-vr/bringing-pixels-front-and-center-vr-video/