feat: vendor threejs files (#247)
Copy the example files we depended on from an older version of threejs into a vendor directory and apply the rollup replace pieces directly to it.
This allows us to update to a newer version of threejs so that npm audit --production no longer produces any issues. We can't upgrade to newer versions of threejs until we update our tooling and likely start transpiling threejs as well, because by default it starts including new JS features in the build files.

Fixes #244.